Reorganize the package.
