+++ .gitattributes
@@ -1,1 +1,1 @@
-bids/reports/_version.py export-subst
+bids/ext/reports/_version.py export-subst
