+++ MANIFEST.in
@@ -1,4 +1,4 @@
 include versioneer.py
-include bids/reports/_version.py
-recursive-include bids/reports/config *.json
-recursive-include bids/reports/tests/data *
+include bids/ext/reports/_version.py
+recursive-include bids/ext/reports/config *.json
+recursive-include bids/ext/reports/tests/data *
