+++ bids/ext/reports/parameters.py
