+++ bids/ext/reports/parsing.py
