+++ setup.cfg
@@ -58,8 +58,8 @@ dev =
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = bids/reports/_version.py
-versionfile_build = bids/reports/_version.py
+versionfile_source = bids/ext/reports/_version.py
+versionfile_build = bids/ext/reports/_version.py
 tag_prefix =
 parentdir_prefix =
